Rename offer external.rs to testutils.rs
