+++ single_offer/src/testutils.rs
